py/requirements.txt
about-time==4.2.1
alabaster==0.7.13
alive-progress==3.1.5
altair==5.2.0
annotated-types==0.6.0
attrs==23.1.0
Babel==2.14.0
bcrypt==4.1.2
black==22.10.0
blinker==1.7.0
brazilcep==6.0.0
cachetools==5.3.2
certifi==2023.5.7
cffi @ file:///private/tmp/cffi-20231015-5890-1t5aans/cffi-1.16.0
charset-normalizer==3.1.0
click==8.1.7
coverage==7.3.4
cryptography @ file:///private/tmp/python-cryptography-20231024-6561-iqijf3/cryptography-41.0.5
cx-Freeze==6.15.11
DateTime==5.4
docopt @ file:///private/tmp/python-docopt-20231106-5487-1a8cp1l/docopt-0.6.2
docutils==0.20.1
Faker==21.0.0
flake8==6.1.0
gitdb==4.0.11
GitPython==3.1.40
grapheme==0.6.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.11.0
iniconfig==2.0.0
invoke==2.2.0
isodate==0.6.1
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keyring==24.3.0
load-dotenv==0.1.0
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.1.0
mypy==1.4.1
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
nassl==5.1.0
nh3==0.2.15
numpy==1.25.0
packaging @ file:///private/tmp/python-packaging-20231018-5901-huh5q4/packaging-23.2
pandas==2.0.2
paramiko==3.4.0
pathspec==0.12.1
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.7.0
pluggy==1.3.0
protobuf==3.20.3
pyarrow==14.0.2
pycodestyle==2.11.1
pycparser @ file:///private/tmp/pycparser-20231013-5647-12y5m5s/pycparser-2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pyflakes==3.1.0
Pygments==2.17.2
PyNaCl==1.5.0
pyOpenSSL==23.3.0
pyperclip==1.8.2
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML @ file:///private/tmp/pyyaml-20231019-6030-1l8rk6l/PyYAML-6.0.1
readme-renderer==42.0
referencing==0.32.0
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
rpds-py==0.16.2
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-autodoc-typehints==1.25.2
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
streamlit==1.29.0
tenacity==8.2.3
tk==0.1.0
tls_parser==2.0.1
toml==0.10.2
toolz==0.12.0
tornado==6.4
twine==4.0.2
types-pyOpenSSL==23.3.0.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.3
validators==0.22.0
zeep==4.2.1
zipp==3.17.0
zope.interface==6.1